backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.7.0
python-multipart==0.0.18
aiofiles==24.1.0
python-dotenv==1.0.1
openai-whisper==20240930
fugashi[unidic-lite]==1.5.2
google-genai==1.0.0
numpy==1.26.4
